UIService/requirements.txt
streamlit
requests
pyjwt
